Add theme component
